+++ packages/react-behave/src/components/Dropdown/Dropdown.test.js
@@ -6,18 +6,16 @@ import { mount } from 'enzyme';
 describe('<Dropdown />', () => {
   it('passes down the ref to `render`', () => {
     const render = jest.fn();
-    mount(<Dropdown render={render} renderDropDown={jest.fn()} />);
+    mount(<Dropdown renderDropDown={jest.fn()}>{render}</Dropdown>);
     expect(render).toHaveBeenCalledWith(expect.anything());
   });
 
   it('passes down the ref and props to `renderDropDown`', () => {
     const renderDropDown = jest.fn(() => <p />);
     mount(
-      <Dropdown
-        open={true}
-        render={jest.fn()}
-        renderDropDown={renderDropDown}
-      />,
+      <Dropdown open={true} renderDropDown={renderDropDown}>
+        {jest.fn()}
+      </Dropdown>,
     );
     expect(renderDropDown).toHaveBeenCalledWith(
       expect.anything(),
@@ -27,18 +25,16 @@ describe('<Dropdown />', () => {
 
   it("doesn't call `renderDropDown` when not opened", () => {
     const renderDropDown = jest.fn();
-    mount(<Dropdown render={jest.fn()} renderDropDown={renderDropDown} />);
+    mount(<Dropdown renderDropDown={renderDropDown}>{jest.fn()}</Dropdown>);
     expect(renderDropDown).not.toHaveBeenCalled();
   });
 
   it("shouldn't add a <ClickOutside /> component if `onClickOutside` is not provided", () => {
     const renderDropDown = jest.fn(() => <p />);
     const wrapper = mount(
-      <Dropdown
-        open={true}
-        render={jest.fn()}
-        renderDropDown={renderDropDown}
-      />,
+      <Dropdown open={true} renderDropDown={renderDropDown}>
+        {jest.fn()}
+      </Dropdown>,
     );
     expect(wrapper.find(ClickOutside)).toHaveLength(0);
   });
@@ -49,9 +45,10 @@ describe('<Dropdown />', () => {
       <Dropdown
         onClickOutside={jest.fn()}
         open={true}
-        render={jest.fn()}
         renderDropDown={renderDropDown}
-      />,
+      >
+        {jest.fn()}
+      </Dropdown>,
     );
     expect(wrapper.find(ClickOutside).first()).toBeDefined();
   });