+++ packages/react-behave/src/components/MergeRefs/MergeRefs.test.js
@@ -7,10 +7,9 @@ describe('<MergeRefs />', () => {
     const refObject = React.createRef();
     const refCallback = jest.fn();
     mount(
-      <MergeRefs
-        refs={[refObject, refCallback]}
-        render={ref => <p ref={ref}>Hello</p>}
-      />,
+      <MergeRefs refs={[refObject, refCallback]}>
+        {ref => <p ref={ref}>Hello</p>}
+      </MergeRefs>,
     );
     expect(refObject.current).toBeDefined();
     expect(refCallback).toHaveBeenCalledWith(expect.anything());
@@ -18,13 +17,13 @@ describe('<MergeRefs />', () => {
 
   it("doesn't pass a ref if the array is empty", () => {
     const render = jest.fn(() => <p>Hello</p>);
-    mount(<MergeRefs refs={[]} render={render} />);
+    mount(<MergeRefs refs={[]}>{render}</MergeRefs>);
     expect(render).toHaveBeenCalledWith(null);
   });
 
   it("doesn't pass a ref if the array only contain falsy values", () => {
     const render = jest.fn(() => <p>Hello</p>);
-    mount(<MergeRefs refs={[null, undefined]} render={render} />);
+    mount(<MergeRefs refs={[null, undefined]}>{render}</MergeRefs>);
     expect(render).toHaveBeenCalledWith(null);
   });
 });